code/requirements.txt
absl-py==2.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
bleach==6.2.0
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.116.1
fastjsonschema==2.21.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.59.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.1
idna==3.10
instaloader==4.14.2
ipykernel==6.29.5
ipython==8.12.3
jax==0.6.2
jaxlib==0.6.2
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kagglehub==0.3.12
kiwisolver==1.4.8
lxml==6.0.0
MarkupSafe==3.0.2
mistune==3.1.3
ml_dtypes==0.5.1
mpmath==1.3.0
narwhals==2.0.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy
opt_einsum==3.4.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
pipreqs==0.5.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
pypdfium2==4.30.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docs==0.1.0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scikit-surprise==1.1.4
scipy==1.13.1
sentencepiece==0.2.0
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
soupsieve==2.7
stack-data==0.6.3
starlette==0.47.2
statsmodels==0.14.5
streamlit==1.47.1
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
timm==1.0.19
tinycss2==1.4.0
tokenizers==0.21.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
undetected-chromedriver==3.5.5
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
wsproto==1.2.0
xlsxwriter==3.2.5
yarg==0.1.9 